# Backport var declaration and reference resolving performance improvements