--- a/.github/workflows/build.yml
+++ b/.github/workflows/build.yml
@@ -3,7 +3,7 @@ name: Build
 on:
 
   push:
-    branches: [ main ]
+    branches: [ main, 3.x ]
     paths-ignore:
     - 'doc/**'
     - '**.md'