--- a/Jint/Runtime/Interpreter/Expressions/DestructuringPatternAssignmentExpression.cs
+++ b/Jint/Runtime/Interpreter/Expressions/DestructuringPatternAssignmentExpression.cs
@@ -145,6 +145,7 @@ namespace Jint.Runtime.Interpreter.Expressions
                     {
                         close = true;
                         var reference = GetReferenceFromMember(context, me);
+
                         JsValue value;
                         if (arrayOperations != null)
                         {